feat: cleanup header files

+++ include/data_structures/tokenizer.h
@@ -9,12 +9,12 @@ typedef struct TokenizerDef {
     Array* tokens;
 } Tokenizer;
 
-Tokenizer* create_tokenizer(unsigned int initial_size);
+Tokenizer* create_tokenizer(size_t initial_size);
 void print_tokens(Tokenizer* tokenizer);
 
 Token* consume(Tokenizer* tokenizer);
 bool peek(Tokenizer* tokenizer, Symbol symbol);
-bool peek_ahead(Tokenizer* tokenizer, Symbol symbol, unsigned int offset);
+bool peek_ahead(Tokenizer* tokenizer, Symbol symbol, size_t offset);
 bool peek_type(Tokenizer* tokenizer);
 Token* peek_consume(Tokenizer* tokenizer, Symbol symbol);
